+++ hephaistos/cli.py
@@ -212,7 +212,7 @@ class PatchSubcommand(BaseSubcommand):
         super().__init__(description="patch Hades using Hephaistos", **kwargs)
         self.add_argument('width', type=int, help="display resolution width")
         self.add_argument('height', type=int, help="display resolution height")
-        self.add_argument('-s', '--scaling', default=Scaling.HOR_PLUS,
+        self.add_argument('--scaling', default=Scaling.HOR_PLUS,
             choices=[Scaling.HOR_PLUS.value, Scaling.PIXEL_BASED.value],
             help="scaling type (default: 'hor+')")
         self.add_argument('--hud', default=HUD.EXPAND,