+++ hephaistos/config.py
@@ -21,7 +21,7 @@ class Screen:
 DEFAULT_SCREEN = Screen(1920, 1080)
 
 # Hephaistos constants
-VERSION = 'v1.3.2'
+VERSION = 'v1.4.0'
 LATEST_RELEASE_URL = 'https://github.com/nbusseneau/hephaistos/releases/latest/'
 LATEST_RELEASE_API_URL = 'https://api.github.com/repos/nbusseneau/hephaistos/releases/latest'
 HEPHAISTOS_NAME = 'hephaistos'