+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+-   Remove `-s` CLI shorthand for `--scaling`.
+
 ## [v1.3.2] - 2021-09-12
 
 ### Fixed